rename mod to index, fix tsconfig, merge licenses
