+++ typegate/tests/random/random.ts
@@ -1,7 +1,7 @@
 // Copyright Metatype OÜ, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
 
-import { Policy, t, typegraph } from "@typegraph/sdk/mod.js";
+import { Policy, t, typegraph } from "@typegraph/sdk/index.js";
 import { RandomRuntime } from "@typegraph/sdk/runtimes/random.js";
 
 typegraph("random", (g: any) => {
