+++ typegate/tests/runtimes/graphql/typegraphs/deno/graphql.ts
@@ -1,7 +1,7 @@
 // Copyright Metatype OÜ, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
 
-import { Policy, t, typegraph } from "@typegraph/sdk/mod.js";
+import { Policy, t, typegraph } from "@typegraph/sdk/index.js";
 import { GraphQLRuntime } from "@typegraph/sdk/runtimes/graphql.js";
 import * as effects from "@typegraph/sdk/effects.js";
