+++ typegate/tests/runtimes/python_wasi/python_wasi.ts
@@ -1,7 +1,7 @@
 // Copyright Metatype OÜ, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
 
-import { Policy, t, typegraph } from "@typegraph/sdk/mod.js";
+import { Policy, t, typegraph } from "@typegraph/sdk/index.js";
 import { PythonRuntime } from "@typegraph/sdk/runtimes/python.js";
 import outdent from "outdent";
