+++ typegate/tests/runtimes/temporal/temporal.ts
@@ -1,7 +1,7 @@
 // Copyright Metatype OÜ, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
 
-import { Policy, t, typegraph } from "@typegraph/sdk/mod.js";
+import { Policy, t, typegraph } from "@typegraph/sdk/index.js";
 import { TemporalRuntime } from "@typegraph/sdk/providers/temporal.js";
 
 typegraph("temporal", (g: any) => {
