+++ typegraph/node/package.json
@@ -3,7 +3,6 @@
   "version": "0.3.1",
   "description": "Declarative API development platform. Build serverless backends with zero-trust and less code, no matter where and how your (legacy) systems are.",
   "type": "module",
-  "main": "mod.js",
   "scripts": {
     "build": "pnpm exec tsc && cp -R ./src/gen ./dist/",
     "test": "echo \"Error: no test specified\" && exit 1"
