+++ typegraph/node/src/runtimes/python.ts
@@ -5,7 +5,7 @@ import * as t from "../types.js";
 import { runtimes } from "../wit.js";
 import { Effect } from "../gen/interfaces/metatype-typegraph-runtimes.js";
 import { Materializer, Runtime } from "./mod.js";
-import { fx } from "../mod.js";
+import { fx } from "../index.js";
 
 interface LambdaMat extends Materializer {
   fn: string;
